+++ strangeflix/subscribe/admin.py
@@ -1,8 +1,9 @@
+# importing django modules
 from django.contrib import admin
 from .models import SubscriptionPlan, Subscriptions, SubscriptionAdditionalTransaction
 
-# Register your models here.
 
+# registering models to admin panel
 admin.site.register(SubscriptionPlan)
 admin.site.register(Subscriptions)
 admin.site.register(SubscriptionAdditionalTransaction)