+++ changes.md
@@ -2,6 +2,10 @@
 ## Log your changes/modifications here with date and name in specified format.
 ## DO NOT add new events at last. Recent changes should be at the top.
 
+### By <b>_Harshit Garg_</b> on <b>_October 12, 2020_</b>
+* Added content section in provider dashboard
+* Fixed bugs in add content section.
+
 ### By <b>_Harshit Garg_</b> on <b>_October 10, 2020_</b>
 * Changes in add content section.
 * Video can be uploaded using link or file.